Add benchmarking suite for `eval()` method.
Adds a suite of simple benchmarks focused on the `Polynomial::eval()`
method. The suite covers floating point (`f64`) and integer values
(`i64`), in addition to arbitrary precision arithmetic (`BigRational`).

For each data type there are two helper functions: one that is used to
construct a representative set of polynomials for that data type, and a
second to actually evaluate a set (`array`) of polynomials drawn from
that set. The final "benchmarked function" (and thus time) for each type
is the sum across all evaluating every polynomial in each set for a
representative set of times.